+++ lib/get.dart
@@ -1,10 +1,10 @@
 library get;
 
-export 'src/getroute.dart';
-export 'src/routes.dart';
-export 'src/snack.dart';
-export 'src/bottomsheet.dart';
-export 'src/snack_route.dart';
-export 'src/route_observer.dart';
-export 'src/getroute_cupertino.dart';
-export 'src/transitions_type.dart';
+export 'src/routes/default_route.dart';
+export 'src/get_main.dart';
+export 'src/snackbar/snack.dart';
+export 'src/bottomsheet/bottomsheet.dart';
+export 'src/snackbar/snack_route.dart';
+export 'src/routes/observers/route_observer.dart';
+export 'src/routes/transitions_type.dart';
+export 'src/platform/platform.dart';