+++ lib/src/platform/platform_io.dart
@@ -1,0 +1,11 @@
+import 'dart:io';
+
+class GeneralPlatform {
+  static bool get isWeb => false;
+  static bool get isMacOS => Platform.isMacOS;
+  static bool get isWindows => Platform.isWindows;
+  static bool get isLinux => Platform.isLinux;
+  static bool get isAndroid => Platform.isAndroid;
+  static bool get isIOS => Platform.isIOS;
+  static bool get isFuchsia => Platform.isFuchsia;
+}