+++ lib/src/platform/platform_web.dart
@@ -1,0 +1,9 @@
+class GeneralPlatform {
+  static bool get isWeb => true;
+  static bool get isMacOS => false;
+  static bool get isWindows => false;
+  static bool get isLinux => false;
+  static bool get isAndroid => false;
+  static bool get isIOS => false;
+  static bool get isFuchsia => false;
+}